refactor: decouple experienceInfo repository
