+++ src/libs/modules/database/repositories/experienceInfo/experienceInfo.interface.ts
@@ -1,0 +1,3 @@
+export interface ExperienceInfoRepository {}
+
+export const ExperienceInfoRepository = Symbol('ExperienceInfoRepository');
