+++ src/libs/modules/database/repositories/experienceInfo/experienceInfo.repository.ts
@@ -1,10 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { ExperienceInfo, Prisma } from '@prisma/client';
 
-import { PrismaService } from '../prisma.service';
+import { ExperienceInfoRepository } from '@libs/modules/database/repositories/experienceInfo/experienceInfo.interface';
+
+import { PrismaService } from '../../prisma.service';
 
 @Injectable()
-export class ExperienceInfoRepository {
+export class ExperienceInfoRepositoryImpl implements ExperienceInfoRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   public async create(data: Prisma.ExperienceInfoCreateInput): Promise<ExperienceInfo> {
